pyproject.toml
[tool.poetry]
name = "hw13"
version = "0.1.0"
description = ""
authors = ["Gena Shpak <gena_shpak@ukr.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.11"
fastapi = "0.104.1"
alembic = "1.13.0"
sqlalchemy = "2.0.23"
asyncpg = "0.29.0"
uvicorn = "0.24.0.post1"
python-jose = {extras = ["cryptography"], version = "3.3.0"}
pydantic = {extras = ["email"], version = "2.5.2"}
python-multipart = "0.0.6"
libgravatar = "1.0.4"
validate-email = "^1.3"
fastapi-mail = "^1.4.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-dotenv = "^1.0.1"
slowapi = "^0.1.9"
redis = "^5.0.4"
cloudinary = "^1.40.0"
pytest = "^8.2.0"


[tool.poetry.group.dev.dependencies]
sphinx = "^7.3.7"


[tool.poetry.group.test.dependencies]
aiosqlite = "^0.20.0"
pytest-asyncio = "^0.23.6"
httpx = "^0.27.0"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "--doctest-modules"
testpaths = "tests"
pythonpath = "."
filterwarnings = "ignore::DeprecationWarning"